refactor callbacks and extern_c.rs
